requirements.txt
requests
datetime
json
os
kafka-python